Refactor: move window files to under the window directory
