+++ src/flutter/shell/platform/linux_embedded/surface/linuxes_surface.h
@@ -7,7 +7,7 @@
 
 #include <memory>
 
-#include "flutter/shell/platform/linux_embedded/surface/native_window.h"
+#include "flutter/shell/platform/linux_embedded/window/native_window.h"
 
 namespace flutter {
