+++ src/flutter/shell/platform/linux_embedded/window/linuxes_window_drm.h
@@ -12,8 +12,8 @@
 #include <thread>
 
 #include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_drm.h"
-#include "flutter/shell/platform/linux_embedded/surface/native_window_drm.h"
 #include "flutter/shell/platform/linux_embedded/window/linuxes_window.h"
+#include "flutter/shell/platform/linux_embedded/window/native_window_drm.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler.h"
 
 namespace flutter {
