+++ src/flutter/shell/platform/linux_embedded/window/linuxes_window_wayland.h
@@ -13,8 +13,8 @@
 #include <vector>
 
 #include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_wayland.h"
-#include "flutter/shell/platform/linux_embedded/surface/native_window_wayland.h"
 #include "flutter/shell/platform/linux_embedded/window/linuxes_window.h"
+#include "flutter/shell/platform/linux_embedded/window/native_window_wayland.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler.h"
 
 extern "C" {
