+++ src/flutter/shell/platform/linux_embedded/window/linuxes_window_x11.h
@@ -8,8 +8,8 @@
 #include <memory>
 
 #include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_x11.h"
-#include "flutter/shell/platform/linux_embedded/surface/native_window_x11.h"
 #include "flutter/shell/platform/linux_embedded/window/linuxes_window.h"
+#include "flutter/shell/platform/linux_embedded/window/native_window_x11.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler.h"
 
 namespace flutter {
